+++ team_integration_test.go
@@ -414,8 +414,6 @@ func TestTeamsUpdateRunTasks(t *testing.T) {
 
 func TestTeamsUpdateManageProjects(t *testing.T) {
 	skipIfFreeOnly(t)
-	skipUnlessBeta(t)
-	skipIfEnterprise(t)
 
 	client := testClient(t)
 	ctx := context.Background()